backend/requirements.txt
fastapi>=0.95.0,<0.96.0
uvicorn>=0.22.0,<0.23.0
pydantic>=1.10.7,<2.0.0
sqlalchemy>=2.0.12,<2.1.0
alembic>=1.10.4,<1.11.0
pymongo>=4.3.3,<4.4.0
python-dotenv>=1.0.0,<1.1.0
passlib>=1.7.4,<1.8.0
python-jose>=3.3.0,<3.4.0
python-multipart>=0.0.6,<0.1.0
httpx>=0.24.0,<0.25.0
pytest>=7.3.1,<7.4.0
pytest-asyncio>=0.21.0,<0.22.0
mypy>=1.2.0,<1.3.0
black>=23.3.0,<23.4.0
isort>=5.12.0,<5.13.0
flake8>=6.0.0,<6.1.0
